+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add support autofocus function for `imx708` sensor.
+
 ## 2.8.0 (2024-11-18)
 
 - Add support for the `imx708` sensor to `UnicamIspCapture`. Note: Autofocus is not supported.